BUG PdfGraphics2D: fixed StackOverflowError by calling PdfGraphics2D#translate(int,int) method (since f73467cc0eadafb43617ddd7f6d9c21d0cae4396 commit).

+++ openpdf/src/main/java/com/lowagie/text/pdf/PdfGraphics2D.java
@@ -810,7 +810,7 @@ public class PdfGraphics2D extends Graphics2D {
      * @see Graphics#translate(int, int)
      */
     public void translate(int x, int y) {
-        translate(x, y);
+        translate((double)x, (double)y);
     }
     
     /**
